refactor: move connection management to base
Removing everything from the connections v1 manager that wasn't specific
to the connections protocol itself.
